Updating

+++ java/java-impl/src/com/intellij/codeInsight/generation/OverrideImplementUtil.java
@@ -25,11 +25,14 @@ import com.intellij.ide.fileTemplates.FileTemplateManager;
 import com.intellij.ide.fileTemplates.FileTemplateUtil;
 import com.intellij.ide.fileTemplates.JavaTemplateUtil;
 import com.intellij.ide.util.MemberChooser;
+import com.intellij.idea.ActionsBundle;
 import com.intellij.lang.java.JavaLanguage;
 import com.intellij.openapi.actionSystem.KeyboardShortcut;
+import com.intellij.openapi.actionSystem.Presentation;
 import com.intellij.openapi.actionSystem.Shortcut;
 import com.intellij.openapi.application.ApplicationManager;
 import com.intellij.openapi.application.Result;
+import com.intellij.openapi.command.CommandProcessor;
 import com.intellij.openapi.command.WriteCommandAction;
 import com.intellij.openapi.diagnostic.Logger;
 import com.intellij.openapi.editor.Editor;
@@ -354,7 +357,7 @@ public class OverrideImplementUtil extends OverrideImplementExploreUtil {
   public static void setupMethodBody(PsiMethod result, PsiMethod originalMethod, PsiClass targetClass) throws IncorrectOperationException {
     boolean isAbstract = originalMethod.hasModifierProperty(PsiModifier.ABSTRACT) || originalMethod.hasModifierProperty(PsiModifier.DEFAULT);
     String templateName = isAbstract ? JavaTemplateUtil.TEMPLATE_IMPLEMENTED_METHOD_BODY : JavaTemplateUtil.TEMPLATE_OVERRIDDEN_METHOD_BODY;
-    FileTemplate template = FileTemplateManager.getInstance().getCodeTemplate(templateName);
+    FileTemplate template = FileTemplateManager.getInstance(originalMethod.getProject()).getCodeTemplate(templateName);
     setupMethodBody(result, originalMethod, targetClass, template);
   }
 
@@ -363,15 +366,20 @@ public class OverrideImplementUtil extends OverrideImplementExploreUtil {
                                      final PsiClass targetClass,
                                      final FileTemplate template) throws IncorrectOperationException {
     if (targetClass.isInterface()) {
-      final PsiCodeBlock body = result.getBody();
-      if (body != null) body.delete();
+      if (isImplementInterfaceInJava8Interface(targetClass)) {
+        PsiUtil.setModifierProperty(result, PsiModifier.DEFAULT, true);
+      }
+      else {
+        final PsiCodeBlock body = result.getBody();
+        if (body != null) body.delete();
+      }
     }
     FileType fileType = FileTypeManager.getInstance().getFileTypeByExtension(template.getExtension());
     PsiType returnType = result.getReturnType();
     if (returnType == null) {
       returnType = PsiType.VOID;
     }
-    Properties properties = FileTemplateManager.getInstance().getDefaultProperties(targetClass.getProject());
+    Properties properties = FileTemplateManager.getInstance(targetClass.getProject()).getDefaultProperties();
     properties.setProperty(FileTemplate.ATTRIBUTE_RETURN_TYPE, returnType.getPresentableText());
     properties.setProperty(FileTemplate.ATTRIBUTE_DEFAULT_RETURN_VALUE, PsiTypesUtil.getDefaultValueOfType(returnType));
     properties.setProperty(FileTemplate.ATTRIBUTE_CALL_SUPER, callSuper(originalMethod, result));
@@ -408,6 +416,16 @@ public class OverrideImplementUtil extends OverrideImplementExploreUtil {
         oldBody.replace(m.getBody());
       }
     }
+  }
+
+  private static boolean isImplementInterfaceInJava8Interface(PsiClass targetClass) {
+    if (!PsiUtil.isLanguageLevel8OrHigher(targetClass)){
+      return false;
+    }
+    final String implementMethodsName = ActionsBundle.message("action.ImplementMethods.text");
+    final Presentation presentation = new Presentation();
+    presentation.setText(implementMethodsName);
+    return presentation.getText().equals(CommandProcessor.getInstance().getCurrentCommandName());
   }
 
   public static void chooseAndOverrideMethods(Project project, Editor editor, PsiClass aClass){
